+++ test/main.dart
@@ -2,25 +2,26 @@
 // source code is governed by a BSD-style license that can be found in the
 // LICENSE file.
 
-import 'package:flutter_test/flutter_test.dart';
 import 'package:hijri/umm_alqura_calendar.dart';
+import 'package:test/test.dart';
 
 void main() {
   UmmAlquraCalendar _hijriDate = new UmmAlquraCalendar();
   _hijriDate.hYear = 1439;
   _hijriDate.hMonth = 10;
   _hijriDate.hDay = 30;
+
   // _hijriDate.currentLocale = 'ar';
   group('Hijri', () {
     test('produces the correct date', () {
-      expect(new UmmAlquraCalendar.fromDate(new DateTime(2018, 5, 27)).toList(),
-          equals([1439, 9, 12]));
+      expect(new UmmAlquraCalendar.fromDate(new DateTime(2020, 5, 20)).toList(),
+          equals([1441, 9, 27]));
     });
     test('format date', () {
       expect(
           new UmmAlquraCalendar.fromDate(new DateTime(2018, 5, 27))
               .toFormat("dd mm yy"),
-          equals("12 09 39"));
+          equals("12 9 39"));
     });
     test('is valid date', () {
       expect(_hijriDate.isValid(), equals(false));