+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog
+## 1.0.1
+
+- add days short name for Arabic
+- Edit local method thanks @emawyas
+
+## 1.0.0
+
+- Change local
 
 ## 0.2.1